refactored out PostItAnalysis
